Fix: moved email verify endpoint to authControler from accountController
