+++ src/main/java/com/mudda/backend/auth/AuthController.java
@@ -9,14 +9,12 @@
 package com.mudda.backend.auth;
 
 import com.mudda.backend.account.AccountService;
+import com.mudda.backend.account.VerifyRequest;
 import com.mudda.backend.user.CreateUserRequest;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/auth")
@@ -35,6 +33,18 @@ public class AuthController {
     public ResponseEntity<String> registerUser(@Valid @RequestBody CreateUserRequest registrationRequest) {
         accountService.register(registrationRequest);
         return ResponseEntity.status(HttpStatus.CREATED).body("Registration successful.");
+    }
+
+    @PostMapping("/verify-email")
+    public ResponseEntity<String> requestVerification(@RequestBody VerifyRequest verifyRequest) {
+        accountService.sendEmailVerificationLink(verifyRequest.email());
+        return ResponseEntity.ok("If account exists verification link has been sent to email.");
+    }
+
+    @GetMapping("/verify-email")
+    public ResponseEntity<String> verifyEmail(@RequestParam String verifyToken) {
+        accountService.verifyEmail(verifyToken);
+        return ResponseEntity.ok("Email verified successfully.");
     }
 
     //    Only login when both token expires
